requirements.txt
streamlit
requests
beautifulsoup4
python-dotenv
ollama
pandas
lxml
urllib3
streamlit-chat